Project_YoungPeopleSurvey
Young people survey is an interesting collection of data where people were asked questions on different categories and they answered these questions on a range of 1 to 5 based on their likings.

Our main objective was to analyze difference in their areas of interests and likings based on gender and location.

After doing the analysis, we can incorporate the findings into Recommender systems.